# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HRMIS API Documentation",
        "description": "Human Resource Management Information System API",
        "contact": {
            "name": "HRMIS Support",
            "email": "admin@hrmis.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/leaves": {
            "get": {
                "tags": [
                    "Leave Management"
                ],
                "summary": "Get all leave requests",
                "description": "Retrieve a list of all leave requests with user information",
                "operationId": "getLeaves",
                "responses": {
                    "200": {
                        "description": "List of leave requests retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "start_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2025-08-15"
                                            },
                                            "end_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2025-08-20"
                                            },
                                            "reason": {
                                                "type": "string",
                                                "example": "Annual vacation"
                                            },
                                            "status": {
                                                "type": "string",
                                                "enum": [
                                                    "Pending",
                                                    "Approved",
                                                    "Rejected"
                                                ],
                                                "example": "Pending"
                                            },
                                            "requested_by": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "requested_by_user": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "john@example.com"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing authentication token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Leave Management"
                ],
                "summary": "Create a new leave request",
                "description": "Submit a new leave request for the authenticated user",
                "operationId": "createLeave",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "start_date",
                                    "end_date",
                                    "reason"
                                ],
                                "properties": {
                                    "start_date": {
                                        "description": "Start date of leave",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-08-15"
                                    },
                                    "end_date": {
                                        "description": "End date of leave",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-08-20"
                                    },
                                    "reason": {
                                        "description": "Reason for leave request",
                                        "type": "string",
                                        "example": "Annual vacation"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Leave request created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "start_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-08-15"
                                        },
                                        "end_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-08-20"
                                        },
                                        "reason": {
                                            "type": "string",
                                            "example": "Annual vacation"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "Pending"
                                        },
                                        "requested_by": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing authentication token"
                    },
                    "422": {
                        "description": "Validation error - Invalid input data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/leaves/user/{id}": {
            "get": {
                "tags": [
                    "Leave Management"
                ],
                "summary": "Get leaves by user ID",
                "description": "Retrieve all leave requests for a specific user",
                "operationId": "getLeavesByUserId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User leaves retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "start_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2025-08-15"
                                            },
                                            "end_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2025-08-20"
                                            },
                                            "reason": {
                                                "type": "string",
                                                "example": "Annual vacation"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "Pending"
                                            },
                                            "requested_by": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing authentication token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}